Create a base class for the models (#1385)
Updated on 2025-02-11

The objective of this PR is to create a base class all models would
inherit from. This base class defines the interface for models and
includes a `__call__` method that allows models to be called with a
prompt and output type directly following the suggestion in #1359 .

Formatting methods for the input and output types are implemented in a
separate class inheriting from the base class `ModelTypeAdapter`. Each
model must set a `type_adapter` attribute in its initialization. The
rationale for separating those from the model class is that the accepted
types for each model and the formatting logic are quite different
concerns compared to the specifics of the model's implementation.
Separating them also allows us to have an easier to read documentation
of the accepted types for each model.

Another change proposed in this PR is to set a mandatory `model_type`
attribute in each model class. This would allow us not to have to rely
on a static list of models through the `APIModel` and `LocalModel`
classes.

The use of the base model class is implemented for the models that have
already been refactored: OpenAI, Gemini, Anthropic and Llamacpp.

---------

Co-authored-by: Robin Picard <robinpicard@device-36.home>